Zer0LLM_all/wandb/requirements.txt
# 核心依赖
wandb>=0.15.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
pyyaml>=6.0
tqdm>=4.62.0

# 交互式脚本依赖
inquirer>=2.10.0
colorama>=0.4.4

# 可选依赖
scikit-learn>=1.0.0  # 用于分析
plotly>=5.5.0        # 用于高级可视化
torch>=1.10.0        # 如果使用PyTorch模型
transformers>=4.15.0 # 如果使用Hugging Face模型 